[Infrastructure] Splitted the data service into multiple chunks
