+++ lib/domain/services/persistence/base_data_service.dart
@@ -1,0 +1,5 @@
+abstract class BaseDataService {
+  Future<void> init();
+
+  Future<void> deleteThemAll();
+}
