+++ lib/infrastructure/infrastructure.dart
@@ -1,6 +1,5 @@
 export 'package:shiori/infrastructure/calculator_service.dart';
 export 'package:shiori/infrastructure/changelog_provider.dart';
-export 'package:shiori/infrastructure/data_service.dart';
 export 'package:shiori/infrastructure/device_info_service.dart';
 export 'package:shiori/infrastructure/game_codes_service.dart';
 export 'package:shiori/infrastructure/genshin_service.dart';
@@ -8,7 +7,7 @@ export 'package:shiori/infrastructure/locale_service.dart';
 export 'package:shiori/infrastructure/logging_service.dart';
 export 'package:shiori/infrastructure/network_service.dart';
 export 'package:shiori/infrastructure/notification_service.dart';
-export 'package:shiori/infrastructure/persistence/custom_builds_data_service.dart';
+export 'package:shiori/infrastructure/persistence/data_service.dart';
 export 'package:shiori/infrastructure/purchase_service.dart';
 export 'package:shiori/infrastructure/settings_service.dart';
 export 'package:shiori/infrastructure/telemetry/telemetry_service.dart';
